+++ src/examples/gipc/counter/sessionport/twopartyconsensus/asymmetric/AGreetingConsensusListener.java
@@ -6,14 +6,14 @@ import consensus.ProposalState;
 public class AGreetingConsensusListener implements ConsensusListener<String>{
 
 	@Override
-	public void newLocalProposalState(int aProposalNumber, String aProposal,
+	public void newLocalProposalState(float aProposalNumber, String aProposal,
 			ProposalState aProposalState) {
 		
 		
 	}
 
 	@Override
-	public void newRemoteProposalState(int aProposalNumber, String aProposal,
+	public void newRemoteProposalState(float aProposalNumber, String aProposal,
 			ProposalState aProposalState) {
 		
 	}
